solating receive perf counter into separate feature
To allow the external package to disable it
