+++ src/NServiceBus.Core/Performance/ReceiveStatisticsFeature.cs
@@ -1,6 +1,5 @@
 ﻿namespace NServiceBus
 {
-    using System.Threading.Tasks;
     using Features;
 
     class ReceiveStatisticsFeature : Feature
@@ -12,36 +11,8 @@
 
         protected internal override void Setup(FeatureConfigurationContext context)
         {
-            var logicalAddress = context.Settings.LogicalAddress();
-            var performanceDiagnosticsBehavior = new ReceivePerformanceDiagnosticsBehavior(logicalAddress.EndpointInstance.Endpoint);
-
-            context.Pipeline.Register(performanceDiagnosticsBehavior, "Provides various performance counters for receive statistics");
             context.Pipeline.Register("ProcessingStatistics", new ProcessingStatisticsBehavior(), "Collects timing for ProcessingStarted and adds the state to determine ProcessingEnded");
             context.Pipeline.Register("AuditProcessingStatistics", new AuditProcessingStatisticsBehavior(), "Add ProcessingStarted and ProcessingEnded headers");
-
-            context.RegisterStartupTask(new WarmupCooldownTask(performanceDiagnosticsBehavior));
-        }
-
-        class WarmupCooldownTask : FeatureStartupTask
-        {
-            public WarmupCooldownTask(ReceivePerformanceDiagnosticsBehavior behavior)
-            {
-                this.behavior = behavior;
-            }
-
-            protected override Task OnStart(IMessageSession session)
-            {
-                behavior.Warmup();
-                return TaskEx.CompletedTask;
-            }
-
-            protected override Task OnStop(IMessageSession session)
-            {
-                behavior.Cooldown();
-                return TaskEx.CompletedTask;
-            }
-
-            readonly ReceivePerformanceDiagnosticsBehavior behavior;
         }
     }
 }
